Fix Lost Particle w/ Runtime Attr
For some processes and/or diagnostics, we add extra runtime
attributes to our beam (particle container). The logic that
collected "lost" particles in the beamline, i.e., as marked as
lost in apertures, did not account for extra runtime attributes
and thus got lost in bookkeeping.

This fixes the collection logic to be more robust and also copy
any extra runtime attributes over to the "lost" particle recording.

This was first seen with an input that used the nonlinear lens (NLL),
NLL-invariant diagnostics and an aperture at the same time.