Dashboard: Improve UI for Inputs (#855)
* use dictionary for consistent padding

* Add card sizing to __main__ rather than in individual cards

This is part of the goal of having the Inputs take up 50% of the page only, leaving the right 50% for future visualization additions.

* Remove 'no_gutters=True' in __main__

 It does not look like it is needed

* style fixup

* Re-order sections and fix reference to padding dictionary and lattice card name

* Place v_show in __main__

So that the line of code is not hidden deeply and can easily read from the 'outside'

* Strip all hard-coded styles from input parameters section

* Add standardized style to rows in input parameters section

* Adjust style for checkboxes

Need to assign 'auto' instead as we want checkboxes to be next to each other.. this is primarily making space for future PR to allow space for 'select mode' select field

* Add UIDefaults class, and standardized styles for the section cards

This is to ensure that the input section can function on a smaller screen

* Update distribution section with similar styling as input parameters card

* Update styling for custom select components

Does not require to 'show_details' as there it is not expected for a dropdown to have validation

* Update csr section with similar styling

* Update space_charge section with similar styling

* update Input __init__

* Update Input/components.py

* update lattice config ui

* Add ToolbarDefaults to defaults.py

* Update UIdefaults and refactor styling to inherit form UIdefaults

* Give cards dynamic sizing based on viewport height

* Improve v_alert's responsiveness

Making it not expand past the toolbar

* add breakpoints for cards/inputs section

* Add custom css for scrollbars

* Make icons into buttons

To improve responsiveness and user expectations

* Update lattice config section

Alter input_header function to allow more flexibility on button placement, update space_charge header because of this. Additionally, the lattice config is inside one vcard rather than two separate, and the icons are treated as actual buttons with an icon.

* remove unused style

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Add staticmethod tag

* Remove scrollbars on inputs

* Add expand button to all sections by default

* Add collapsable sections

* Add collapse all button in toolbar

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>