Examples: Default openPMD Backend
We default to ADIOS2 output when available for
performance in the I/O of our `BeamMonitor` element.

For CI and correctness tests, because sometimes it is easier/faster
to install from packages (e.g., Debian), we hard-code the HDF5
openPMD backend, to avoid that the test behaves differently if
ADIOS2 is present (switches to bp) or not (switches to h5).

Now, we can actually let the code switch automatically because
with openPMD 0.16+, we have a generic pattern `.%E` in analsy
scripts to read the written files and match the first openPMD
backend that works (by file name extension). That way, we can now
simplify our examples by not adding the extra lines to control
the openPMD backend anymore.