Tests: `amrex.the_arena_init_size=0`
`amrex.the_arena_init_size=0` is default on CPU but not GPU.
In order to run tests in parallel on the same GPU, we need to
change the default so the first test does not allocate 3/4th
of the memory of the GPU (to speed up later allocations in
a pre-allocated heap).

This currently only applys to app tests and pytest, but not
yet to the Python CTests. Need to add an extra option for those,
because we cannot inject CLI options.