solve reactivity issues
The UI will now properly update to user expectations. For example if a user enters a non-existing variable name first, then enters the variable name in the config and a valid value, the lattice list config will automatically interpret it. Previously, this same case would have required the user to reenter the var name (ie. AFTER the variable was made in the config)