Dashboard: Simulation Progress (#917)
* Use an async subprocess to run the simulation

* Move location, rename, and make 'exportTemplate' file the main simulation

Now we do not have to update two separate files when adding in a new dashboard input

* Add sim timer

To see how long a simulation takes

* Add step tracker

* Update 'run simulation' button on ui

* rename: variables

* Add sim progress

* update: sim logic

* add reset

Called as the first execution when the run sim is clicked

Resets everything sim related

* Add sim_progress_status

shows up under the progress bar

* update: run_sim button

* Allow subprocess to generate PS plots

* simplify/organize code

* Remove interactive phase space plot

As discussed, the current interactive mode is complex to facilitate when the simulation is ran on a subprocess. a future pr may re-implement the phase space interactive modes in plotly, which is more trame friendly.

* Remove simulation.py and helper functions

Now that simulations run directly from the exported script via a subprocess, we don't need separate simulation logic or duplicated helper functions. This way, there's only one source of truth—the same script users get when exporting their inputs. It's simpler and easier to maintain, even though we're now building simulations through f-strings.

* Rename files

* Modify the sim script to not display the generating phase space code when exporting inputs

* rename sim function

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fix: codeQL

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>