Explicit Power 2 & 3
Alex reminded me, that pow is really slow, e.g., on GPU.
After a few experiments with Godbolt, it seems that indeed
modern compilers do not replace powers of 2 and 3 with a
(fast) multiplication unless fast-math is used. We currently
only use fast-math on GPU, not CPU.

Thus, simple instances of `std::pow(..., 2)` and `std::pow(..., 3)`
are replaced now with explict multiplications.

In many cases, writing power ^2 and ^3 is also shorter without
`std::pow`.