PipeAperture Mix-In w/o Divisions (#1011)
* PipeAperture Mixing w/o Divisions

The `PipeAperture` class is a mixin class used in all
thick elements and always checked and potentially applied
for on the critical path of the computation.

A small optimization we can apply is to pre-compute the
inverse of the pipe aperture in x and y, so that we can do
a simple multiplication to calculate aperture horizontal
and vertical comparison instead of a (way more) costly
division.

Co-authored-by: Alexander Sinn <alexander.sinn@desy.de>

* Add Inline Comments

On performance optimization.

---------

Co-authored-by: Alexander Sinn <alexander.sinn@desy.de>