Python: Interactive/In Situ Beam History
Add support to in situ store the reduced beam history with a beam
instead of writing it to file. This is useful for interactive use
of the simulation and to avoid writing files.

Before that, we had access to this information only through
text files and we could access already this information, but
only for the very last step or by writing our own push/tracker
loop in Python.