Dashboard: Restructure Input Files (#1035)
* rename: csr files

* rename: lattice files

and fix imports

* rename: space charge files

* rename: 'DistributionParameters' class -> 'DistributionConfiguration'

All other classes which store inputs end with 'configuration' as well so this is for consistency

* fix: spelling oncsr/isr configuration class name

instead of csrConfiguration, rename to CSRConfiguration. same with isr

this is for proper python convention

* rename: input parameter files

* make imports clearer

* remove duplicate entres from Input/__init__.py

* adjust components import

better to enter the components package first?

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fix import for ci

* fix import

* rename input parameters to simulation parameters across dashbaord

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>