fix: error with state.kin_energy_unit
because the input of state.kin_energy_unit is coming from a dropdown, its input is a string. if it goes through the shared.py state change '@state.change(*INPUT_DEFAULTS)', then it will be validated as a string. Because it has no error messages (technically because the state.kin_energy_unit default is 'MeV' and would be a valid python identifier), it will be converted to a numeric value. however the 'convert_to_numeric' method will set the value to 'None' if it can't be turned into a default type of int or float. this will set the state.kin_energy_unit to 'None' and the dashboard will not launch. The solution to this is not have any dropdown inputs go through @state.change(*INPUT_DEFAULTS) because every dropdown will have a valid set of inputs by default.