fix dashboard test for lattice
The current default for lattice name is 'None', a string. We change that to a python literal None now. But because that is the case the name parameter for the lattice element will be empty. So we need to make sure in the dashboard test we provide a name for the element so that there are no inputs errors before running a simulation