Dashboard: refactor validation (#1051)
* update validate_against's arguments

* refactor validate_against

break into multiple smaller functions

* refactor validation logic in shared.py

* simplify n_cell and blocking_factor validation

* simplify shared.py with usingnon-hardcoded defaults

* cleanup

docstring, type hint, error messages

* prob_relative validation cleanup

* update type hint

* cleanup

docstring/simplify

* move validation function to validation.py

* simplify lattice validation

* fix from rebase

* cleanup

* fix: error with state.kin_energy_unit

because the input of state.kin_energy_unit is coming from a dropdown, its input is a string. if it goes through the shared.py state change '@state.change(*INPUT_DEFAULTS)', then it will be validated as a string. Because it has no error messages (technically because the state.kin_energy_unit default is 'MeV' and would be a valid python identifier), it will be converted to a numeric value. however the 'convert_to_numeric' method will set the value to 'None' if it can't be turned into a default type of int or float. this will set the state.kin_energy_unit to 'None' and the dashboard will not launch. The solution to this is not have any dropdown inputs go through @state.change(*INPUT_DEFAULTS) because every dropdown will have a valid set of inputs by default.

* rename INPUT_DEFAULTS -> STATE_INPUTS

* fix convert_to_numeric

which also fixes distribution validaiton

* fix lattice validation and remove print statement

* remove: duplicate import

* remove comment

* fix dashboard test for lattice

The current default for lattice name is 'None', a string. We change that to a python literal None now. But because that is the case the name parameter for the lattice element will be empty. So we need to make sure in the dashboard test we provide a name for the element so that there are no inputs errors before running a simulation

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* rename validate_input to validate

* print what was shown in the terminal page if 'assert_state' reached timeoutError

* update update_distribution_parameter arguments

* cleanup build_lattice

no need to re-validate

the build_lattice function is only called at simulation time, where at that point all values on the dashboard are valid

* add validation folder to Input/

* split logic of validation/inputs.py and validation/errors_tracker.py

* fix validation imports

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>