Copy the Reference Particle (#1122)
* Copy the Reference Particle

Use this if you want to do an operation that modifies the reference
particle, but you want to avoid modifying the particle container
that it is assigned to.

Example: Plotting a lattice by doing a loop over all elements. In
that loop, one wants to do always plot the next element, then push
a reference particle through it, then continue with the next element.
This is important to get the global coordinates of the next element
and to properly treat accelerating elements (e.g., because energy
changes the bending radius `rc` of following elements).

* Test and `push(ref, element)`

* Cleanup: `impactx::push` is a function

Thus needs to be lowercase.