Reduced Beam Characteristics: Rename mean/min/max/sig (#1128)
This renames the components in the Reduced Beam Characteristics
for consistency:
- `x/y/t_mean/min/max` -> `mean/min/max_x/y/t`
- `sig_x/y/t` -> `sigma_x/y/t`

For the text file `diags/reduced_beam_characteristics*`, this is
a breaking change.

For the openPMD attributes and Python API, the old names are now
undocumented and deprecated, but still available. Users are
encouraged to transition to the new names, as the old names will
be removed in a future version of ImpactX.