Dashboard: Fix population of distribution values & Python 3.14 issue (#1208)
* fix population of distribution values

unsure if the mechanisms of state.flush changed on the trame end, but it looks like now it executes only after the function ends and not at the moment that it is called. this 'fix' populates the values on the ui once the state.flush is called.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* try: event loop before server.start()

---------

Co-authored-by: ax3l <1353258+ax3l@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>