Fix: `track_particles` with `1` particle
We can in principle allow to run with `1` particle
in `track_particles()`, even though we do have the
`track_reference(ref)` mode now.

As long as no space charge and no CSR/wakefield effects are
requested, this is still a valid configuration.

This PR removes an old assert, for which we have a more specific
assert in the space charge routines, if enabled. Additionally,
it adds a new check into the `HandleWakefield` routines (for CSR),
which also rely on a non-flat beam to model anything sensible.

Other routines we have, notably ISR, are modeled on individual
particles and need no asserts.